Cast output to float64 in inference
Otherwise numpy conversion fails when using 16 bit precision

+++ kraken/lib/models.py
@@ -109,7 +109,7 @@ class TorchSeqRecognizer(object):
         o, olens = self.nn.nn(line, lens)
         if o.size(2) != 1:
             raise KrakenInputException('Expected dimension 3 to be 1, actual {}'.format(o.size()))
-        self.outputs = o.detach().squeeze(2).cpu().numpy()
+        self.outputs = o.detach().squeeze(2).float().cpu().numpy()
         if olens is not None:
             olens = olens.cpu().numpy()
         return self.outputs, olens
